requirements.txt
alembic==1.6.5
atomicwrites==1.4.0
attrs==21.4.0
blinker==1.4
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
coverage==6.4.1
decorator==5.1.0
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
Flask-WTF==0.15.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.1
infinity==1.5
iniconfig==1.1.1
intervals==0.9.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
packaging==21.3
phonenumbers==8.12.34
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.11.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.22
SQLAlchemy-Utils==0.37.8
tomli==2.0.1
twilio==7.0.0
typing-extensions==3.10.0.2
urllib3==1.26.7
validators==0.18.2
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
WTForms-Alchemy==0.17.0
WTForms-Components==0.10.5
zipp==3.6.0
